Time: 2 ms (54.46%), Space: 6.3 MB (63.34%) - LeetHub

+++ 0790-domino-and-tromino-tiling/0790-domino-and-tromino-tiling.cpp
@@ -5,7 +5,7 @@ public:
         if (n <= 2) return n;
         long fPrev = 1L, fCurr = 2L, pCurr = 1L;
         for (int k = 3; k <= n; ++k){
-            int tmp = fCurr;
+            long tmp = fCurr;
             fCurr = (fPrev + fCurr + 2 * pCurr) % mod;
             pCurr = (pCurr + fPrev) % mod;
             fPrev = tmp;
